* attribute and feature extraction

+++ src/include/crf/features/feature_gen.h
@@ -7,8 +7,8 @@ namespace NLP {
       public:
         FeatureGen(const Type &type) : type(type) { }
         virtual ~FeatureGen(void) { }
-        virtual void operator()(Attributes &attributes, Sentence &sent, TagPair &tp, size_t j) = 0;
-        virtual void operator()(Attributes &attributes, Sentence &sent, Context &c, size_t j) = 0;
+        virtual void operator()(Attributes &attributes, Sentence &sent, TagPair &tp, int j) = 0;
+        virtual void operator()(Attributes &attributes, Sentence &sent, Context &c, int j) = 0;
     };
 
     class WordGen : public FeatureGen {
@@ -16,8 +16,8 @@ namespace NLP {
         WordGen(const Type &type);
         virtual ~WordGen(void) { }
 
-        virtual void operator()(Attributes &attributes, Sentence &sent, TagPair &tp, size_t j);
-        virtual void operator()(Attributes &attributes, Sentence &sent, Context &c, size_t j);
+        virtual void operator()(Attributes &attributes, Sentence &sent, TagPair &tp, int j);
+        virtual void operator()(Attributes &attributes, Sentence &sent, Context &c, int j);
 
     };
   }
